// src/types/investor.ts
export type InvestorData = {
  limited_partner_type: string | null;
  aum: number | null;
  year_founded: number | null;
  hqlocation: string | null;
  hqemail: string | null;
  hqphone: string | null;
  description: string | null;
  preferred_fund_type: string | null;
  preferred_commitment_size_min: number | null;
  preferred_commitment_size_max: number | null;
  open_to_first_time_funds: string | null;
  primary_contact: string | null;
  primary_contact_title: string | null;
  primary_contact_email: string | null;
  primary_contact_phone: string | null;
  website: string | null;
  private_equity: number | null;
  private_equity_percent: number | null;
  real_estate: number | null;
  real_estate_percent: number | null;
  special_opportunities: number | null;
  special_opportunities_percent: number | null;
  hedge_funds: number | null;
  hedge_funds_percent: number | null;
  equities: number | null;
  equities_percent: number | null;
  fixed_income: number | null;
  fixed_income_percent: number | null;
  cash: number | null;
  cash_percent: number | null;
};